backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-genai==1.5.0
gunicorn==22.0.0
whitenoise==6.6.0
dj-database-url==2.2.0
